Refactor Page and RawPage
The IPage protocol is getting a handler method meant to give the
current handler for the page. This should allow for a cleaner
injection of middleware into any type of page. Currently RawPage has a
working implementation, but without any testing against what currently
exists.

As a step towards addding working middleware for pages the with-access
macro has been removed from the RawPage render method. A more flexible
middleware solution is to be developed instead.