Add wrap-exeptions middleware
The authn and authz handling in reverie, as well as error handling,
suffers from an old understand of best practices. Authn and authz will
be updated keeping the old approach, while at the same time adding
a middleware based approach that will be used for defpage and
defmodule.

Usage of slingshot has also been removed in this implementation as
ex-info works just as well, and is supported out of the box in later
versions of Clojure.

Wrap-exceptions handles any exceptions, with a default exception
handler used if a specific type can't be found.

Important information about the exceptions are logged as WARN, and is
not returned in the response.