Refactor routing
Routing was tied up into a complex, but contained mess, inside of
site.clj. With this refactoring the routing has been broken out of
Site into a new namespace router under http. A clearer distinction
has also been made between routes that are coming from the database,
and routes that are static (RawPage and Module).

In addition, some functionality was deeply tied into Site and has been
refactored into using the middlware pattern. They are in order: CSRF
protection (and if it should be used or not), caching and handling of
responses. These end up in a new middleware namespace for site.

And further, the middleware functionality specific for RawPage (as of
this point in time), is also taken care of by the page-handler; also
situated in the new namespace for site.

Finally, Site opens up for multiple routes from the database, allowing
adding multiple hosts in the reverie database. This will probably have
to be revisted when multiple hosts are introduced into reverie and its
database schema.