Add styling for reverie-admin
Ant Design, while having a fantastic library of components, is very
opinionated about how the styling should be done. Unlike in Bootstrap,
Compass, etc where you can easily build on what's been defined before,
Ant Design makes this approach very hard to do. Doubly so with
Clojurescript.

Because of this a general build aimed at just styling is done, and any
other styling that would need to be done is done in a separate
stylesheet instead.

Also, Ant Design is styled in LESS, while SASS was picked for the
second stylesheet. Main reason is that Ant Design's choices in styling
makes things very complicated, and for reverie's usage, too
complicated. No real advantage is gained, and so the easier and
simpler path has been chosen instead.