Add model for objects
Only added get-objects, get-available-objects and get-datasource for
now. Moving to a SPA type of admin interface means untangling a number
of things that was done in the previous interface. In this case we
need to separate objects that are available in a page, the meta for
the objects and additional datasources for the meta (such as options
for :dropdown fields).