Fix routing and update IMigrator protocol
Routing spec was faulty, since the :reverie.http.route/casting spec
only took regex into account. However, if you decided to define the
regex in a var and reference that, it then becomes a symbol in the
macros, thus invalidating the spec. Updated the spec to take this into
account.

The IMigrator protocol has been extended to give a fair amount of more
details and control over migrations. Exposes a number of things in
migratus, which reverie currently uses. The methods added to the
protocol should still hold if another migration library was decided to
be used, as they should be general purpose enough.